requirements.txt
python-telegram-bot[rate-limiter]==21.6
minio==7.2.7
aiohttp==3.9.5
prometheus-client==0.20.0
python-slugify==8.0.4
Unidecode==1.3.8
python-json-logger==2.0.7
uvloop==0.19.0
aiodns==3.2.0
orjson==3.10.7

